made modifications to defuzzifier
- calculation of centroids happens on creation of fuzzy set
- changed crispValue from int to float and cascaded changes
- deleted redundant outputVariables arraylist from FuzzySystem and from Main
- implemented crisp value calculation method for output variables in Defuzzifier

+++ Fuzzifier.java
@@ -1,7 +1,7 @@
 import java.util.ArrayList;
 
 public abstract class Fuzzifier {
-    public static float calculateGraphRule (float AX, float AY, float BX, float BY, int x){
+    public static float calculateGraphRule (float AX, float AY, float BX, float BY, float x){
         float m = (BY - AY) / (BX - AX);
         float c = AY - (m * AX);
         float y = (m * x) + c;
